Flip bits

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int bits = 0, res = n ^ m;
+	unsigned long int bits = 0, res = n ^ m;
 
 	while (res)
 	{
